Refactor FamiliaViewModel and enhance dashboard UI
Refactored FamiliaViewModel to simplify property declarations, improve chart and filter logic, and add support for dynamic Top N selection in dashboard charts. Updated FamiliasView.xaml to enhance dashboard layout, add new card styles, improve data grid and chart presentation, and synchronize filter bindings. Minor updates to MainViewModel and MainWindow.xaml for snackbar support and UI consistency.

+++ MVVM/Models/VentaReporteModel.cs
@@ -21,6 +21,6 @@ namespace WPF_PAR.MVVM.Models
         public decimal Descuento { get; set; }
         public decimal TotalVenta { get; set; }
         public double LitrosUnitarios { get; set; }
-        public double LitrosTotales => Cantidad * LitrosUnitarios;
+        public double LitrosTotales => ( double ) Cantidad * LitrosUnitarios;
     }
 }
